requirements.txt
Django==5.0.6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
sentence-transformers
faiss-cpu
gunicorn
APScheduler
python-dotenv
